refactor(populate): move `getModelsMapForPopulate()` into separate function
Fix #7962
