+++ test/document.populate.test.js
@@ -479,19 +479,15 @@ describe('document.populate', function() {
   });
 
   describe('of new document', function() {
-    it('should save just the populated _id (gh-1442)', function(done) {
+    it('should save just the populated _id (gh-1442)', function() {
       const b = new B({_creator: user1});
-      b.populate('_creator', function(err, b) {
-        if (err) return done(err);
+      return co(function*() {
+        yield b.populate('_creator').execPopulate();
         assert.equal(b._creator.name, 'Phoenix');
-        b.save(function(err) {
-          assert.ifError(err);
-          B.collection.findOne({_id: b._id}, function(err, b) {
-            assert.ifError(err);
-            assert.equal(b._creator, String(user1._id));
-            done();
-          });
-        });
+        yield b.save();
+
+        const _b = yield B.collection.findOne({_id: b._id});
+        assert.equal(_b._creator.toString(), String(user1._id));
       });
     });
   });
